requirements.txt
pymupdf
pdf2image
pytesseract
opencv-python
faiss-cpu
sentence-transformers
openai
numpy
groq
apscheduler
flask
flask-cors
python-dotenv